pyproject.toml
[project]
name = "enz-rxn-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "stefanpate", email = "stefanpate94@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ergochemics>=0.1.2",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=6.29.5",
    "lxml>=5.3.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "polars>=1.31.0",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "rdkit>=2024.9.5",
    "requests>=2.32.3",
    "setuptools>=80.7.1",
    "tqdm>=4.67.1",
]

[project.scripts]
enz-rxn-data = "enz_rxn_data:main"

[project.optional-dependencies]
thermo = [
    "equilibrator-api>=0.6.0",
    "equilibrator-assets>=0.6.0",
    "torch~=2.10.0",
    "torch-geometric~=2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
